ingestion/requirements.txt
fastapi
uvicorn
python-dotenv
redis
slowapi
prometheus-fastapi-instrumentator
aiofiles
pytest
